llm_svs_automation/requirements.txt
numpy
pandas
matplotlib
seaborn
tqdm
librosa
soundfile
pyyaml
resemblyzer
